Double -> BigDecimal netto barnetilsyn

+++ src/main/java/no/nav/bidrag/beregn/nettobarnetilsyn/beregning/NettoBarnetilsynBeregning.java
@@ -1,5 +1,6 @@
 package no.nav.bidrag.beregn.nettobarnetilsyn.beregning;
 
+import java.math.BigDecimal;
 import java.util.List;
 import no.nav.bidrag.beregn.nettobarnetilsyn.bo.GrunnlagBeregningPeriodisert;
 import no.nav.bidrag.beregn.nettobarnetilsyn.bo.ResultatBeregning;
@@ -9,10 +10,10 @@ public interface NettoBarnetilsynBeregning {
   List<ResultatBeregning> beregn(
       GrunnlagBeregningPeriodisert grunnlagBeregningPeriodisert);
 
-  Double beregnFradragsbelopPerBarn(
+  BigDecimal beregnFradragsbelopPerBarn(
       GrunnlagBeregningPeriodisert grunnlagBeregningPeriodisert,
       int antallBarn, int antallBarnMedTilsynsutgift,
-      double tilsynsbelop);
+      BigDecimal tilsynsbelop);
 
   static NettoBarnetilsynBeregning getInstance(){
     return new NettoBarnetilsynBeregningImpl();
